CRUD IMAGENES A PRUEBA DE ERRORES
Esta aplicacion web de prueba para subir imagenes esta configurado correctamente para:
- Subir una imagen en una carpeta assets que se va a mostrar en la aplicacion web.
- Evitar que un usuario envie el formulario con campos vacios y/o con un archivo que no sea de tipo imagen.
- Editar un registro de tal manera que se pueda conservar la imagen original.
- Si el usuario elimina definitivamente un registro, entonces tambien se va a eliminar la imagen.

+++ InitializrSpringbootProject/src/main/java/com/prueba01/service/GaleriaServiceImpl.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+
 @Service
 public class GaleriaServiceImpl implements GaleriaService {
 
@@ -107,6 +108,13 @@ public GaleriaDto updateRegistro(GaleriaDto registro) {
         return null;
     }
 
+    /*
+    // Si el archivo subido no es de tipo image, entonces no se va a guardar el registro.
+    if(!registro.getImagen().getContentType().startsWith("image/")){
+        return null;
+    }
+*/
+    
     // Si la imagen no está vacía y es diferente a la imagen actual, eliminar la imagen anterior y almacenar la nueva
     if (registro.getImagen() != null && !registro.getImagen().isEmpty() && !registro.getImagen().equals(galeriaEntity.getImagen())) {
         contenedorImagenesImpl.eliminarImagen(galeriaEntity.getRutaImagen());
